refactoring categorize each API by router
